app/streamlit/requirements.txt
requests==2.31.0
streamlit==1.39.0
aiohttp==3.10.10
supabase==2.9.1
plotly==5.19.0
python-dotenv==1.0.1
